Abstract dates utility into dates file. Add react timezone
