+++ package.json
@@ -40,6 +40,7 @@
     "react-router-hash-link": "^1.2.0",
     "react-scripts": "2.1.5",
     "react-select": "^2.1.0",
+    "react-timezone": "^2.3.0",
     "react-tooltip": "^3.9.0",
     "truffle": "^5.0.0-beta.0",
     "uuid": "^3.3.2",
